+++ modules/textadept/bookmarks.lua
@@ -5,7 +5,7 @@
 local M = {}
 
 --- The bookmark mark number.
-M.MARK_BOOKMARK = _SCINTILLA.new_marker_number()
+M.MARK_BOOKMARK = view.new_marker_number()
 
 --- Toggles a bookmark on the current line.
 function M.toggle()