+++ modules/textadept/editing.lua
@@ -40,7 +40,7 @@ M.auto_enclose = false
 M.typeover_auto_paired = true
 
 --- The word highlight indicator number.
-M.INDIC_HIGHLIGHT = _SCINTILLA.new_indic_number()
+M.INDIC_HIGHLIGHT = view.new_indic_number()
 
 --- Map of image names to registered image numbers.
 -- @field CLASS The image number for classes.
@@ -62,7 +62,7 @@ events.connect(events.VIEW_NEW, function()
 		if type(name) == 'string' then view:register_image(i, M.XPM_IMAGES[i]) end
 	end
 end)
-for _ = 1, #M.XPM_IMAGES do _SCINTILLA.new_image_type() end -- sync
+for _ = 1, #M.XPM_IMAGES do view.new_image_type() end -- sync
 
 --- Map of lexer names to line comment strings for programming languages, used by
 -- `editing.toggle_comment()`.