+++ modules/textadept/run.lua
@@ -19,13 +19,13 @@ M.run_in_background = false
 M.run_without_prompt = false
 
 --- The run or compile warning marker number.
-M.MARK_WARNING = _SCINTILLA.new_marker_number()
+M.MARK_WARNING = view.new_marker_number()
 --- The run or compile error marker number.
-M.MARK_ERROR = _SCINTILLA.new_marker_number()
+M.MARK_ERROR = view.new_marker_number()
 --- The run or compile warning indicator number.
-M.INDIC_WARNING = _SCINTILLA.new_indic_number()
+M.INDIC_WARNING = view.new_indic_number()
 --- The run or compile error indicator number.
-M.INDIC_ERROR = _SCINTILLA.new_indic_number()
+M.INDIC_ERROR = view.new_indic_number()
 
 -- Events.
 local run_events = {'compile_output', 'run_output', 'build_output', 'test_output'}